requirements.txt
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.10.2
dash-table==5.0.0
faiss-cpu==1.7.2
filelock==3.8.0
Flask==2.2.2
Flask-Compress==1.12
fonttools==4.36.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.0
transformers==4.21.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
